pyproject.toml
[project]
name = "mv-people"
version = "0.1.0"
description = "A CLI tool to scan folders for images of people and archive them."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "opencv-python-headless>=4.0.0",
    "numpy>=1.20.0",
    "pillow>=9.0.0",
]

[project.scripts]
mv-people = "mv_people.main:scan"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pyinstaller>=6.0.0",
]
